==> Revise_Experiment/Revisi 3/1.SplitDataset.ipynb <==
# import library
import pandas as pd 
import numpy as np 
import matplotlib.pyplot as plt 
import seaborn as sns 
import plotly.express as px 
from datetime import datetime
import os
from sklearn.model_selection import train_test_split
%matplotlib inline

# tabel setting
pd.set_option('display.max_columns',None)
pd.set_option('display.max_rows',None)
pd.set_option('display.precision', 2)
pd.options.display.float_format = '{:,.2f}'.format

dataset_dir = '../../Dataset/'

sensor1_path = dataset_dir + 'NCC-2 Dataset Simultaneous Botnet Dataset/sensor1/'
os.listdir(sensor1_path)

sensor2_path = dataset_dir + 'NCC-2 Dataset Simultaneous Botnet Dataset/sensor2/'
os.listdir(sensor2_path)

sensor3_path = dataset_dir + 'NCC-2 Dataset Simultaneous Botnet Dataset/sensor3/'
os.listdir(sensor3_path)

# Split Data Train n Test

def split_data(df):
    train, test = train_test_split(df, test_size=0.25, random_state=42)
    return train, test

# Sensor 1

df_1 = pd.read_csv(sensor1_path + 'sensor1.binetflow')

sensor1_train, sensor1_test = split_data(df_1)

sensor1_train.head()

sensor1_train.shape

sensor1_test.head()

sensor1_test.shape

sensor1_test.to_csv(dataset_dir + 'sensor1_test.csv', index=False)
sensor1_train.to_csv(dataset_dir + 'sensor1_train.csv', index=False)

# Sensor 2

df_2 = pd.read_csv(sensor2_path + 'sensor2.binetflow')

sensor2_train, sensor2_test = split_data(df_2)

sensor2_train.head()

sensor2_train.shape

sensor2_test.head()

sensor2_test.shape

sensor2_test.to_csv(dataset_dir + 'sensor2_test.csv', index=False)
sensor2_train.to_csv(dataset_dir + 'sensor2_train.csv', index=False)

# Sensor 3

df_3 = pd.read_csv(sensor3_path + 'sensor3.binetflow')

sensor3_train, sensor3_test = split_data(df_3)

sensor3_train.head()

sensor3_train.shape

sensor3_test.head()

sensor3_test.shape

sensor3_test.to_csv(dataset_dir + 'sensor3_test.csv', index=False)
sensor3_train.to_csv(dataset_dir + 'sensor3_train.csv', index=False)
