==> Revise_Experiment/Revisi 3/2a.PreProcessing_Sensor1 copy.ipynb <==
# import library
import pandas as pd 
import numpy as np 
import matplotlib.pyplot as plt 
import seaborn as sns 
import plotly.express as px 
from datetime import datetime
import os
%matplotlib inline

# tabel setting
# pd.set_option('display.max_columns',None)
# pd.set_option('display.max_rows',None)
# pd.set_option('display.precision', 2)
# pd.options.display.float_format = '{:,.2f}'.format

# Import Dataset

dataset_dir = '../../Dataset/'

# Labeling

def categorize_label(label):
    label = label.lower()  
    if 'botnet' in label and 'spam' in label:
        return 2
    elif 'botnet' in label:
        return 1
    else:
        return 0

# Sensor 1

sensor1_train = pd.read_csv(dataset_dir + 'sensor1_train.csv')
sensor1_test = pd.read_csv(dataset_dir + 'sensor1_test.csv')

sensor1_train = sensor1_train.drop(columns=['ActivityLabel', 'BotnetName', 'SensorId'])
sensor1_test = sensor1_test.drop(columns=['ActivityLabel', 'BotnetName', 'SensorId'])

sensor1_train['Label'] = sensor1_train['Label'].apply(categorize_label)
sensor1_test['Label'] = sensor1_test['Label'].apply(categorize_label)

sensor1_train.sample(10)

sensor1_test.sample(10)

sensor1_train['Label'].value_counts()

sensor1_test['Label'].value_counts()

## Cleaning Data

### Drop Duplicates

sensor1_train = sensor1_train.drop_duplicates()
sensor1_test = sensor1_test.drop_duplicates()

print(sensor1_train.shape)
print(sensor1_test.shape)

### Drop dTos

#### Train

cek = sensor1_train[(sensor1_train['dTos'].isnull())]

cek['Label'].value_counts()

cek = cek[cek['Label'] == 0]

cek['Label'].value_counts()

sensor1_train = sensor1_train.drop(cek.index)

#### Test

cek = sensor1_test[(sensor1_test['dTos'].isnull())]

cek['Label'].value_counts()

cek = cek[cek['Label'] == 0]

cek['Label'].value_counts()

sensor1_test = sensor1_test.drop(cek.index)

### Drop sTos

#### Train

cek = sensor1_train[(sensor1_train['sTos'].isnull())]

cek['Label'].value_counts()

sensor1_train = sensor1_train.drop(cek.index)

#### Test

cek = sensor1_test[(sensor1_test['sTos'].isnull())]

cek['Label'].value_counts()

sensor1_test = sensor1_test.drop(cek.index)

### Drop Dport

#### Train

cek = sensor1_train[sensor1_train['Dport'].isnull()]

cek['Label'].value_counts()

cek = cek[cek['Label'] == 0]

cek['Label'].value_counts()

sensor1_train = sensor1_train.drop(cek.index)

#### Test

cek = sensor1_test[sensor1_test['Dport'].isnull()]

cek['Label'].value_counts()

cek = cek[cek['Label'] == 0]

cek['Label'].value_counts()

sensor1_test = sensor1_test.drop(cek.index)

### Drop Sport

#### Train 

cek = sensor1_train[sensor1_train['Sport'].isnull()]

cek['Label'].value_counts()

#### Test

cek = sensor1_test[sensor1_test['Sport'].isnull()]

cek['Label'].value_counts()

### Drop State

#### Train

cek = sensor1_train[sensor1_train['State'].isnull()]

cek['Label'].value_counts()

#### Test

cek = sensor1_test[sensor1_test['State'].isnull()]

cek['Label'].value_counts()

## Imputing

### dTos

sensor1_train['dTos'].value_counts()

cek = sensor1_train[sensor1_train['dTos'].isnull()]

cek['Label'].value_counts()

# Impute with modes
mode_value1 = sensor1_train[sensor1_train['Label'] == 1]['dTos'].mode()[0]
mode_value2 = sensor1_train[sensor1_train['Label'] == 2]['dTos'].mode()[0]

sensor1_train.loc[(sensor1_train['dTos'].isnull()) & (sensor1_train['Label'] == 1), 'dTos'] = mode_value1
sensor1_train.loc[(sensor1_train['dTos'].isnull()) & (sensor1_train['Label'] == 2), 'dTos'] = mode_value2

sensor1_test.loc[(sensor1_test['dTos'].isnull()) & (sensor1_test['Label'] == 1), 'dTos'] = mode_value1
sensor1_test.loc[(sensor1_test['dTos'].isnull()) & (sensor1_test['Label'] == 2), 'dTos'] = mode_value2

# Cek Missing Data
total_missing = sensor1_train.isnull().sum().sort_values(ascending=False)
percent_1 = sensor1_train.isnull().sum()/sensor1_train.isnull().count()*100
percent_2 = (round(percent_1, 1)).sort_values(ascending=False)
missing_data = pd.concat([total_missing, percent_2], axis=1, keys=['Total Missing', '%'])
missing_data = missing_data.reset_index().rename(columns={'index': 'Column'})
missing_data

# Cek Missing Data
total_missing = sensor1_test.isnull().sum().sort_values(ascending=False)
percent_1 = sensor1_test.isnull().sum()/sensor1_test.isnull().count()*100
percent_2 = (round(percent_1, 1)).sort_values(ascending=False)
missing_data = pd.concat([total_missing, percent_2], axis=1, keys=['Total Missing', '%'])
missing_data = missing_data.reset_index().rename(columns={'index': 'Column'})
missing_data

### Dport

sensor1_train['Dport'].value_counts()

cek = sensor1_train[sensor1_train['Dport'].isnull()]

cek['Label'].value_counts()

cek = sensor1_test[sensor1_test['Dport'].isnull()]

cek['Label'].value_counts()

# Impute with modes
mode_value = sensor1_train[sensor1_train['Label'] == 1]['Dport'].mode()[0]

sensor1_train['Dport'] = sensor1_train['Dport'].fillna(mode_value)

sensor1_test['Dport'] = sensor1_test['Dport'].fillna(mode_value)

### Sport

sensor1_train['Sport'].value_counts()

cek = sensor1_train[sensor1_train['Sport'].isnull()]

cek['Label'].value_counts()

cek = sensor1_test[sensor1_test['Sport'].isnull()]

cek['Label'].value_counts()

# Impute with modes
mode_value = sensor1_train[sensor1_train['Label'] == 1]['Sport'].mode()[0]

sensor1_train['Sport'] = sensor1_train['Sport'].fillna(mode_value)

sensor1_test['Sport'] = sensor1_test['Sport'].fillna(mode_value)

### State

sensor1_train['State'].value_counts()

cek = sensor1_train[sensor1_train['State'].isnull()]

cek['Label'].value_counts()

cek = sensor1_test[sensor1_test['State'].isnull()]

cek['Label'].value_counts()

# Impute with modes
mode_value = sensor1_train[sensor1_train['Label'] == 1]['State'].mode()[0]

sensor1_train['State'] = sensor1_train['State'].fillna(mode_value)
sensor1_test['State'] = sensor1_test['State'].fillna(mode_value)

# Cek Missing Data
total_missing = sensor1_train.isnull().sum().sort_values(ascending=False)
percent_1 = sensor1_train.isnull().sum()/sensor1_train.isnull().count()*100
percent_2 = (round(percent_1, 1)).sort_values(ascending=False)
missing_data = pd.concat([total_missing, percent_2], axis=1, keys=['Total Missing', '%'])
missing_data = missing_data.reset_index().rename(columns={'index': 'Column'})
missing_data

# Cek Missing Data
total_missing = sensor1_test.isnull().sum().sort_values(ascending=False)
percent_1 = sensor1_test.isnull().sum()/sensor1_test.isnull().count()*100
percent_2 = (round(percent_1, 1)).sort_values(ascending=False)
missing_data = pd.concat([total_missing, percent_2], axis=1, keys=['Total Missing', '%'])
missing_data = missing_data.reset_index().rename(columns={'index': 'Column'})
missing_data

sensor1_train.to_csv(dataset_dir + 'train1_clean.csv', index=False)

sensor1_test.to_csv(dataset_dir + 'test1_clean.csv', index=False)
