==> Revise_Experiment/Revisi 3/2b.PreProcessing_Sensor2.ipynb <==
# import library
import pandas as pd 
import numpy as np 
import matplotlib.pyplot as plt 
import seaborn as sns 
import plotly.express as px 
from datetime import datetime
import os
%matplotlib inline

# tabel setting
# pd.set_option('display.max_columns',None)
# pd.set_option('display.max_rows',None)
# pd.set_option('display.precision', 2)
# pd.options.display.float_format = '{:,.2f}'.format

# Import Dataset

dataset_dir = '../../Dataset/'

# Labeling

def categorize_label(label):
    label = label.lower()  
    if 'botnet' in label and 'spam' in label:
        return 2
    elif 'botnet' in label:
        return 1
    else:
        return 0

# Sensor 2

sensor2_train = pd.read_csv(dataset_dir + 'sensor2_train.csv')
sensor2_test = pd.read_csv(dataset_dir + 'sensor2_test.csv')

sensor2_train = sensor2_train.drop(columns=['ActivityLabel', 'BotnetName', 'SensorId'])
sensor2_test = sensor2_test.drop(columns=['ActivityLabel', 'BotnetName', 'SensorId'])

sensor2_train['Label'] = sensor2_train['Label'].apply(categorize_label)
sensor2_test['Label'] = sensor2_test['Label'].apply(categorize_label)

sensor2_train.sample(10)

sensor2_test.sample(10)

sensor2_train['Label'].value_counts()

sensor2_test['Label'].value_counts()

## Cleaning Data

### Drop Duplicates

sensor2_train = sensor2_train.drop_duplicates()
sensor2_test = sensor2_test.drop_duplicates()

print(sensor2_train.shape)
print(sensor2_test.shape)

### Drop dTos

#### Train

cek = sensor2_train[(sensor2_train['dTos'].isnull())]

cek['Label'].value_counts()

cek = cek[cek['Label'] == 0]

cek['Label'].value_counts()

sensor2_train = sensor2_train.drop(cek.index)

#### Test

cek = sensor2_test[(sensor2_test['dTos'].isnull())]

cek['Label'].value_counts()

cek = cek[cek['Label'] == 0]

cek['Label'].value_counts()

sensor2_test = sensor2_test.drop(cek.index)

### Drop sTos

#### Train

cek = sensor2_train[(sensor2_train['sTos'].isnull())]

cek['Label'].value_counts()

sensor2_train = sensor2_train.drop(cek.index)

#### Test

cek = sensor2_test[(sensor2_test['sTos'].isnull())]

cek['Label'].value_counts()

sensor2_test = sensor2_test.drop(cek.index)

### Drop Dport

#### Train

cek = sensor2_train[sensor2_train['Dport'].isnull()]

cek['Label'].value_counts()

cek = cek[cek['Label'] == 0]

cek['Label'].value_counts()

sensor2_train = sensor2_train.drop(cek.index)

#### Test

cek = sensor2_test[sensor2_test['Dport'].isnull()]

cek['Label'].value_counts()

cek = cek[cek['Label'] == 0]

cek['Label'].value_counts()

sensor2_test = sensor2_test.drop(cek.index)

### Drop Sport

#### Train 

cek = sensor2_train[sensor2_train['Sport'].isnull()]

cek['Label'].value_counts()

#### Test

cek = sensor2_test[sensor2_test['Sport'].isnull()]

cek['Label'].value_counts()

### Drop State

#### Train

cek = sensor2_train[sensor2_train['State'].isnull()]

cek['Label'].value_counts()

#### Test

cek = sensor2_test[sensor2_test['State'].isnull()]

cek['Label'].value_counts()

## Imputing

### dTos

sensor2_train['dTos'].value_counts()

cek = sensor2_train[sensor2_train['dTos'].isnull()]

cek['Label'].value_counts()

# Impute with modes
mode_value1 = sensor2_train[sensor2_train['Label'] == 1]['dTos'].mode()[0]
mode_value2 = sensor2_train[sensor2_train['Label'] == 2]['dTos'].mode()[0]

sensor2_train.loc[(sensor2_train['dTos'].isnull()) & (sensor2_train['Label'] == 1), 'dTos'] = mode_value1
sensor2_train.loc[(sensor2_train['dTos'].isnull()) & (sensor2_train['Label'] == 2), 'dTos'] = mode_value2

sensor2_test.loc[(sensor2_test['dTos'].isnull()) & (sensor2_test['Label'] == 1), 'dTos'] = mode_value1
sensor2_test.loc[(sensor2_test['dTos'].isnull()) & (sensor2_test['Label'] == 2), 'dTos'] = mode_value2

# Cek Missing Data
total_missing = sensor2_train.isnull().sum().sort_values(ascending=False)
percent_1 = sensor2_train.isnull().sum()/sensor2_train.isnull().count()*100
percent_2 = (round(percent_1, 1)).sort_values(ascending=False)
missing_data = pd.concat([total_missing, percent_2], axis=1, keys=['Total Missing', '%'])
missing_data = missing_data.reset_index().rename(columns={'index': 'Column'})
missing_data

# Cek Missing Data
total_missing = sensor2_test.isnull().sum().sort_values(ascending=False)
percent_1 = sensor2_test.isnull().sum()/sensor2_test.isnull().count()*100
percent_2 = (round(percent_1, 1)).sort_values(ascending=False)
missing_data = pd.concat([total_missing, percent_2], axis=1, keys=['Total Missing', '%'])
missing_data = missing_data.reset_index().rename(columns={'index': 'Column'})
missing_data

### Dport

sensor2_train['Dport'].value_counts()

cek = sensor2_train[sensor2_train['Dport'].isnull()]

cek['Label'].value_counts()

cek = sensor2_test[sensor2_test['Dport'].isnull()]

cek['Label'].value_counts()

# Impute with modes
mode_value = sensor2_train[sensor2_train['Label'] == 1]['Dport'].mode()[0]

sensor2_train['Dport'] = sensor2_train['Dport'].fillna(mode_value)

sensor2_test['Dport'] = sensor2_test['Dport'].fillna(mode_value)

### Sport

sensor2_train['Sport'].value_counts()

cek = sensor2_train[sensor2_train['Sport'].isnull()]

cek['Label'].value_counts()

cek = sensor2_test[sensor2_test['Sport'].isnull()]

cek['Label'].value_counts()

# Impute with modes
mode_value = sensor2_train[sensor2_train['Label'] == 1]['Sport'].mode()[0]

sensor2_train['Sport'] = sensor2_train['Sport'].fillna(mode_value)

sensor2_test['Sport'] = sensor2_test['Sport'].fillna(mode_value)

### State

sensor2_train['State'].value_counts()

cek = sensor2_train[sensor2_train['State'].isnull()]

cek['Label'].value_counts()

cek = sensor2_test[sensor2_test['State'].isnull()]

cek['Label'].value_counts()

# Impute with modes
mode_value = sensor2_train[sensor2_train['Label'] == 1]['State'].mode()[0]

sensor2_train['State'] = sensor2_train['State'].fillna(mode_value)
sensor2_test['State'] = sensor2_test['State'].fillna(mode_value)

# Cek Missing Data
total_missing = sensor2_train.isnull().sum().sort_values(ascending=False)
percent_1 = sensor2_train.isnull().sum()/sensor2_train.isnull().count()*100
percent_2 = (round(percent_1, 1)).sort_values(ascending=False)
missing_data = pd.concat([total_missing, percent_2], axis=1, keys=['Total Missing', '%'])
missing_data = missing_data.reset_index().rename(columns={'index': 'Column'})
missing_data

# Cek Missing Data
total_missing = sensor2_test.isnull().sum().sort_values(ascending=False)
percent_1 = sensor2_test.isnull().sum()/sensor2_test.isnull().count()*100
percent_2 = (round(percent_1, 1)).sort_values(ascending=False)
missing_data = pd.concat([total_missing, percent_2], axis=1, keys=['Total Missing', '%'])
missing_data = missing_data.reset_index().rename(columns={'index': 'Column'})
missing_data

sensor2_train.to_csv(dataset_dir + 'train2_clean.csv', index=False)

sensor2_test.to_csv(dataset_dir + 'test2_clean.csv', index=False)
