==> Revise_Experiment/Revisi 3/2c.PreProcessing_Sensor3.ipynb <==
# import library
import pandas as pd 
import numpy as np 
import matplotlib.pyplot as plt 
import seaborn as sns 
import plotly.express as px 
from datetime import datetime
import os
%matplotlib inline

# tabel setting
# pd.set_option('display.max_columns',None)
# pd.set_option('display.max_rows',None)
# pd.set_option('display.precision', 2)
# pd.options.display.float_format = '{:,.2f}'.format

# Import Dataset

dataset_dir = '../../Dataset/'

# Labeling

def categorize_label(label):
    label = label.lower()  
    if 'botnet' in label and 'spam' in label:
        return 2
    elif 'botnet' in label:
        return 1
    else:
        return 0

# Sensor 3

sensor3_train = pd.read_csv(dataset_dir + 'sensor3_train.csv')
sensor3_test = pd.read_csv(dataset_dir + 'sensor3_test.csv')

sensor3_train = sensor3_train.drop(columns=['ActivityLabel', 'BotnetName', 'SensorId'])
sensor3_test = sensor3_test.drop(columns=['ActivityLabel', 'BotnetName', 'SensorId'])

sensor3_train['Label'] = sensor3_train['Label'].apply(categorize_label)
sensor3_test['Label'] = sensor3_test['Label'].apply(categorize_label)

sensor3_train.sample(10)

sensor3_test.sample(10)

sensor3_train['Label'].value_counts()

sensor3_test['Label'].value_counts()

## Cleaning Data

### Drop Duplicates

sensor3_train = sensor3_train.drop_duplicates()
sensor3_test = sensor3_test.drop_duplicates()

print(sensor3_train.shape)
print(sensor3_test.shape)

### Drop dTos

#### Train

cek = sensor3_train[(sensor3_train['dTos'].isnull())]

cek['Label'].value_counts()

cek = cek[cek['Label'] == 0]

cek['Label'].value_counts()

sensor3_train = sensor3_train.drop(cek.index)

#### Test

cek = sensor3_test[(sensor3_test['dTos'].isnull())]

cek['Label'].value_counts()

cek = cek[cek['Label'] == 0]

cek['Label'].value_counts()

sensor3_test = sensor3_test.drop(cek.index)

### Drop sTos

#### Train

cek = sensor3_train[(sensor3_train['sTos'].isnull())]

cek['Label'].value_counts()

sensor3_train = sensor3_train.drop(cek.index)

#### Test

cek = sensor3_test[(sensor3_test['sTos'].isnull())]

cek['Label'].value_counts()

sensor3_test = sensor3_test.drop(cek.index)

### Drop Dport

#### Train

cek = sensor3_train[sensor3_train['Dport'].isnull()]

cek['Label'].value_counts()

cek = cek[cek['Label'] == 0]

cek['Label'].value_counts()

sensor3_train = sensor3_train.drop(cek.index)

#### Test

cek = sensor3_test[sensor3_test['Dport'].isnull()]

cek['Label'].value_counts()

cek = cek[cek['Label'] == 0]

cek['Label'].value_counts()

sensor3_test = sensor3_test.drop(cek.index)

### Drop Sport

#### Train 

cek = sensor3_train[sensor3_train['Sport'].isnull()]

cek['Label'].value_counts()

cek = cek[cek['Label'] == 0]

cek['Label'].value_counts()

sensor3_train = sensor3_train.drop(cek.index)

#### Test

cek = sensor3_test[sensor3_test['Sport'].isnull()]

cek['Label'].value_counts()

cek = cek[cek['Label'] == 0]

cek['Label'].value_counts()

sensor3_test = sensor3_test.drop(cek.index)

### Drop State

#### Train

cek = sensor3_train[sensor3_train['State'].isnull()]

cek['Label'].value_counts()

#### Test

cek = sensor3_test[sensor3_test['State'].isnull()]

cek['Label'].value_counts()

## Imputing

### dTos

sensor3_train['dTos'].value_counts()

cek = sensor3_train[sensor3_train['dTos'].isnull()]

cek['Label'].value_counts()

# Impute with modes
mode_value1 = sensor3_train[sensor3_train['Label'] == 1]['dTos'].mode()[0]
mode_value2 = sensor3_train[sensor3_train['Label'] == 2]['dTos'].mode()[0]

sensor3_train.loc[(sensor3_train['dTos'].isnull()) & (sensor3_train['Label'] == 1), 'dTos'] = mode_value1
sensor3_train.loc[(sensor3_train['dTos'].isnull()) & (sensor3_train['Label'] == 2), 'dTos'] = mode_value2

sensor3_test.loc[(sensor3_test['dTos'].isnull()) & (sensor3_test['Label'] == 1), 'dTos'] = mode_value1
sensor3_test.loc[(sensor3_test['dTos'].isnull()) & (sensor3_test['Label'] == 2), 'dTos'] = mode_value2

# Cek Missing Data
total_missing = sensor3_train.isnull().sum().sort_values(ascending=False)
percent_1 = sensor3_train.isnull().sum()/sensor3_train.isnull().count()*100
percent_2 = (round(percent_1, 1)).sort_values(ascending=False)
missing_data = pd.concat([total_missing, percent_2], axis=1, keys=['Total Missing', '%'])
missing_data = missing_data.reset_index().rename(columns={'index': 'Column'})
missing_data

# Cek Missing Data
total_missing = sensor3_test.isnull().sum().sort_values(ascending=False)
percent_1 = sensor3_test.isnull().sum()/sensor3_test.isnull().count()*100
percent_2 = (round(percent_1, 1)).sort_values(ascending=False)
missing_data = pd.concat([total_missing, percent_2], axis=1, keys=['Total Missing', '%'])
missing_data = missing_data.reset_index().rename(columns={'index': 'Column'})
missing_data

### Dport

sensor3_train['Dport'].value_counts()

cek = sensor3_train[sensor3_train['Dport'].isnull()]

cek['Label'].value_counts()

cek = sensor3_test[sensor3_test['Dport'].isnull()]

cek['Label'].value_counts()

# Impute with modes
mode_value = sensor3_train[sensor3_train['Label'] == 1]['Dport'].mode()[0]

sensor3_train['Dport'] = sensor3_train['Dport'].fillna(mode_value)

sensor3_test['Dport'] = sensor3_test['Dport'].fillna(mode_value)

### Sport

sensor3_train['Sport'].value_counts()

cek = sensor3_train[sensor3_train['Sport'].isnull()]

cek['Label'].value_counts()

cek = sensor3_test[sensor3_test['Sport'].isnull()]

cek['Label'].value_counts()

# Impute with modes
mode_value = sensor3_train[sensor3_train['Label'] == 1]['Sport'].mode()[0]

sensor3_train['Sport'] = sensor3_train['Sport'].fillna(mode_value)

sensor3_test['Sport'] = sensor3_test['Sport'].fillna(mode_value)

### State

sensor3_train['State'].value_counts()

cek = sensor3_train[sensor3_train['State'].isnull()]

cek['Label'].value_counts()

cek = sensor3_test[sensor3_test['State'].isnull()]

cek['Label'].value_counts()

# Impute with modes
mode_value = sensor3_train[sensor3_train['Label'] == 1]['State'].mode()[0]

sensor3_train['State'] = sensor3_train['State'].fillna(mode_value)
sensor3_test['State'] = sensor3_test['State'].fillna(mode_value)

# Cek Missing Data
total_missing = sensor3_train.isnull().sum().sort_values(ascending=False)
percent_1 = sensor3_train.isnull().sum()/sensor3_train.isnull().count()*100
percent_2 = (round(percent_1, 1)).sort_values(ascending=False)
missing_data = pd.concat([total_missing, percent_2], axis=1, keys=['Total Missing', '%'])
missing_data = missing_data.reset_index().rename(columns={'index': 'Column'})
missing_data

# Cek Missing Data
total_missing = sensor3_test.isnull().sum().sort_values(ascending=False)
percent_1 = sensor3_test.isnull().sum()/sensor3_test.isnull().count()*100
percent_2 = (round(percent_1, 1)).sort_values(ascending=False)
missing_data = pd.concat([total_missing, percent_2], axis=1, keys=['Total Missing', '%'])
missing_data = missing_data.reset_index().rename(columns={'index': 'Column'})
missing_data

sensor3_train.to_csv(dataset_dir + 'train3_clean.csv', index=False)

sensor3_test.to_csv(dataset_dir + 'test3_clean.csv', index=False)
