==> Revise_Experiment/Revisi 3/3.EncodeData.ipynb <==
# import library
import pandas as pd 
import numpy as np 
import matplotlib.pyplot as plt 
import seaborn as sns 
import plotly.express as px 
from datetime import datetime
import os
from sklearn.model_selection import train_test_split
%matplotlib inline

# tabel setting
pd.set_option('display.max_columns',None)
pd.set_option('display.max_rows',None)
pd.set_option('display.precision', 2)
pd.options.display.float_format = '{:,.2f}'.format

dataset_dir = '../../Dataset/'

os.listdir(dataset_dir)

# Encode Categorical

# Frequency Encoding
def frequency_mapping(df):
    freq_map = df.value_counts().to_dict()  # Create frequency map
    
    return freq_map

from sklearn.preprocessing import LabelEncoder
def encode(df, srcaddr_map, dstaddr_map, sport_map, dport_map):
    # Start Time
    df['StartTime'] = pd.to_datetime(df['StartTime'], errors='coerce')
    df['StartTimeHour'] = df['StartTime'].dt.hour
    df['StartTimeMinute'] = df['StartTime'].dt.minute
    df['StartTimeSecond'] = df['StartTime'].dt.second
    df = df.drop(columns=['StartTime'])

    # SrcAddr & DstAddr & Sport & Dport Freq Encoding
    df['SrcAddr'] = df['SrcAddr'].map(lambda x: srcaddr_map.get(x, 0))
    df['DstAddr'] = df['DstAddr'].map(lambda x: dstaddr_map.get(x, 0))
    df['Sport'] = df['Sport'].map(lambda x: sport_map.get(x, 0))
    df['Dport'] = df['Dport'].map(lambda x: dport_map.get(x, 0))

    # Dir & State & dTos & sTos & Proto Label Encoding
    le = LabelEncoder()
    df['Dir'] = le.fit_transform(df['Dir'].astype(str))
    df['State'] = le.fit_transform(df['State'].astype(str))
    df['sTos'] = le.fit_transform(df['sTos'].astype(str))
    df['dTos'] = le.fit_transform(df['dTos'].astype(str))
    df['Proto'] = le.fit_transform(df['Proto'].astype(str))

    return df

def add_target(df):
    df['isBotnet'] = df['Label'].apply(lambda x: 1 if x == 1 or x == 2 else 0)
    df['isSpam'] = df['Label'].apply(lambda x: 1 if x == 2 else 0)
    return df

# Sensor 1

sensor1_train = pd.read_csv(dataset_dir + 'train1_clean.csv')
sensor1_test = pd.read_csv(dataset_dir + 'test1_clean.csv')

sensor1_train.head()

sensor1_test.head()

sensor1_srcaddr_map = frequency_mapping(sensor1_train['SrcAddr'])
sensor1_dstaddr_map = frequency_mapping(sensor1_train['DstAddr'])
sensor1_sport_map = frequency_mapping(sensor1_train['Sport'])
sensor1_dport_map = frequency_mapping(sensor1_train['Dport'])

sensor1_train = encode(sensor1_train, sensor1_srcaddr_map, sensor1_dstaddr_map, sensor1_sport_map, sensor1_dport_map)
sensor1_train = add_target(sensor1_train)

sensor1_test = encode(sensor1_test, sensor1_srcaddr_map, sensor1_dstaddr_map, sensor1_sport_map, sensor1_dport_map)
sensor1_test = add_target(sensor1_test)

sensor1_train.head()

sensor1_test.head()

sensor1_train.to_csv(dataset_dir + 'train1_encoded.csv', index=False)
sensor1_test.to_csv(dataset_dir + 'test1_encoded.csv', index=False)

# Sensor 2

sensor2_train = pd.read_csv(dataset_dir + 'train2_clean.csv')
sensor2_test = pd.read_csv(dataset_dir + 'test2_clean.csv')

sensor2_train.head()

sensor2_test.head()

sensor2_srcaddr_map = frequency_mapping(sensor2_train['SrcAddr'])
sensor2_dstaddr_map = frequency_mapping(sensor2_train['DstAddr'])
sensor2_sport_map = frequency_mapping(sensor2_train['Sport'])
sensor2_dport_map = frequency_mapping(sensor2_train['Dport'])

sensor2_train = encode(sensor2_train, sensor2_srcaddr_map, sensor2_dstaddr_map, sensor2_sport_map, sensor2_dport_map)
sensor2_train = add_target(sensor2_train)

sensor2_test = encode(sensor2_test, sensor2_srcaddr_map, sensor2_dstaddr_map, sensor2_sport_map, sensor2_dport_map)
sensor2_test = add_target(sensor2_test)

sensor2_train.head()

sensor2_test.head()

sensor2_train.to_csv(dataset_dir + 'train2_encoded.csv', index=False)
sensor2_test.to_csv(dataset_dir + 'test2_encoded.csv', index=False)

# Sensor 3

sensor3_train = pd.read_csv(dataset_dir + 'train3_clean.csv')
sensor3_test = pd.read_csv(dataset_dir + 'test3_clean.csv')

sensor3_train.head()

sensor3_test.head()

sensor3_srcaddr_map = frequency_mapping(sensor3_train['SrcAddr'])
sensor3_dstaddr_map = frequency_mapping(sensor3_train['DstAddr'])
sensor3_sport_map = frequency_mapping(sensor3_train['Sport'])
sensor3_dport_map = frequency_mapping(sensor3_train['Dport'])

sensor3_train = encode(sensor3_train, sensor3_srcaddr_map, sensor3_dstaddr_map, sensor3_sport_map, sensor3_dport_map)
sensor3_train = add_target(sensor3_train)

sensor3_test = encode(sensor3_test, sensor3_srcaddr_map, sensor3_dstaddr_map, sensor3_sport_map, sensor3_dport_map)
sensor3_test = add_target(sensor3_test)

sensor3_train.head()

sensor3_test.head()

sensor3_train.to_csv(dataset_dir + 'train3_encoded.csv', index=False)
sensor3_test.to_csv(dataset_dir + 'test3_encoded.csv', index=False)
