==> Revise_Experiment/Revisi 3/4.SplitDataTrain.ipynb <==
# import library
import pandas as pd 
import numpy as np 
import matplotlib.pyplot as plt 
import seaborn as sns 
import plotly.express as px 
from datetime import datetime
import os
from sklearn.model_selection import train_test_split
%matplotlib inline

# tabel setting
pd.set_option('display.max_columns',None)
pd.set_option('display.max_rows',None)
pd.set_option('display.precision', 2)
pd.options.display.float_format = '{:,.2f}'.format

dataset_dir = '../../Dataset/'

# Sensor 1

sensor1_train = pd.read_csv(dataset_dir + 'train1_encoded.csv')

sensor1_train.head()

sensor1_train.shape

## isBotnet

sensor1_isbotnet = sensor1_train.copy()
sensor1_isbotnet.head()

sensor1_isbotnet = sensor1_isbotnet.drop(columns=['Label', 'isSpam'])

sensor1_isbotnet.head()

sensor1_isbotnet.to_csv(dataset_dir + 'sensor1_isbotnet.csv', index=False)

## isSpam

sensor1_isspam = sensor1_train.copy()
sensor1_isspam = sensor1_isspam[sensor1_isspam['isBotnet'] == 1]
sensor1_isspam.head()

sensor1_isspam = sensor1_isspam.drop(columns=['Label', 'isBotnet'])

sensor1_isspam.head()

sensor1_isspam.to_csv(dataset_dir + 'sensor1_isspam.csv', index=False)

# Sensor 2

sensor2_train = pd.read_csv(dataset_dir + 'train2_encoded.csv')

sensor2_train.head()

sensor2_train.shape

## isBotnet

sensor2_isbotnet = sensor2_train.copy()
sensor2_isbotnet.head()

sensor2_isbotnet = sensor2_isbotnet.drop(columns=['Label', 'isSpam'])

sensor2_isbotnet.head()

sensor2_isbotnet.to_csv(dataset_dir + 'sensor2_isbotnet.csv', index=False)

## isSpam

sensor2_isspam = sensor2_train.copy()
sensor2_isspam = sensor2_isspam[sensor2_isspam['isBotnet'] == 1]
sensor2_isspam.head()

sensor2_isspam = sensor2_isspam.drop(columns=['Label', 'isBotnet'])

sensor2_isspam.head()

sensor2_isspam.to_csv(dataset_dir + 'sensor2_isspam.csv', index=False)

# Sensor 3

sensor3_train = pd.read_csv(dataset_dir + 'train3_encoded.csv')

sensor3_train.head()

sensor3_train.shape

## isBotnet

sensor3_isbotnet = sensor3_train.copy()
sensor3_isbotnet.head()

sensor3_isbotnet = sensor3_isbotnet.drop(columns=['Label', 'isSpam'])

sensor3_isbotnet.head()

sensor3_isbotnet.to_csv(dataset_dir + 'sensor3_isbotnet.csv', index=False)

## isSpam

sensor3_isspam = sensor3_train.copy()
sensor3_isspam = sensor3_isspam[sensor3_isspam['isBotnet'] == 1]
sensor3_isspam.head()

sensor3_isspam = sensor3_isspam.drop(columns=['Label', 'isBotnet'])

sensor3_isspam.head()

sensor3_isspam.to_csv(dataset_dir + 'sensor3_isspam.csv', index=False)
